requirements.txt
pytesseract
pydantic
openai
pillow
